useAppList hook wip
